+++ mycodo/mycodo_flask/utils/utils_general.py
@@ -641,15 +641,18 @@ def choices_controller_ids():
     return choices
 
 
-def choices_custom_functions():
+def choices_custom_functions(prepend_func_title=True):
     """ populate form multi-select choices from Function entries """
     choices = []
     dict_controllers = parse_function_information()
     list_controllers_sorted = generate_form_controller_list(dict_controllers)
     for each_custom in list_controllers_sorted:
         value = '{inp}'.format(inp=each_custom)
-        display = 'Function: {name}'.format(
-            name=dict_controllers[each_custom]['function_name'])
+        if prepend_func_title:
+            display = 'Function: {name}'.format(
+                name=dict_controllers[each_custom]['function_name'])
+        else:
+            display = dict_controllers[each_custom]['function_name']
         choices.append({'value': value, 'item': display})
     return choices